Change id type to long

+++ src/main/java/com/frolov/testing/entity/Discipline.java
@@ -16,12 +16,12 @@ public class Discipline extends BaseEntity {
     }
 
     @Override
-    public Integer getId() {
+    public Long getId() {
         return super.getId();
     }
 
     @Override
-    public void setId(Integer id) {
+    public void setId(Long id) {
         super.setId(id);
     }
